linting fixes

+++ packages/windmill/src/services/electoral_log.rs
@@ -692,7 +692,7 @@ impl TryFrom<&Row> for ElectoralLogRow {
     }
 }
 
-pub const IMMUDB_ROWS_LIMIT: u32 = 2500;
+pub const IMMUDB_ROWS_LIMIT: usize = 2500;
 
 #[instrument(err)]
 pub async fn list_electoral_log(input: GetElectoralLogBody) -> Result<DataList<ElectoralLogRow>> {
